# cc-webapp/current_openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Casino-Club F2P API",
    "description": "Backend API for Casino-Club F2P gaming platform",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/segments/user/{user_id}": {
      "get": {
        "tags": [
          "Segments"
        ],
        "summary": "Get User Segment",
        "description": "Get segmentation information for a specific user\nReturns RFM group, LTV score, and risk profile",
        "operationId": "get_user_segment_api_segments_user__user_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserSegmentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/segments/list": {
      "get": {
        "tags": [
          "Segments"
        ],
        "summary": "List All Segments",
        "description": "Get statistics for all user segments\nAdmin only endpoint for business intelligence",
        "operationId": "list_all_segments_api_segments_list_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SegmentStatsResponse"
                  },
                  "type": "array",
                  "title": "Response List All Segments Api Segments List Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/segments/update": {
      "put": {
        "tags": [
          "Segments"
        ],
        "summary": "Update User Segment",
        "description": "Update user segment information\nAdmin only endpoint for manual segment adjustments",
        "operationId": "update_user_segment_api_segments_update_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSegmentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserSegmentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/segments/calculate-rfm": {
      "post": {
        "tags": [
          "Segments"
        ],
        "summary": "Calculate Rfm Segments",
        "description": "Recalculate RFM segments for all users\nAdmin only endpoint for batch segment updates",
        "operationId": "calculate_rfm_segments_api_segments_calculate_rfm_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "days",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "description": "Number of days to analyze for RFM calculation",
              "default": 90,
              "title": "Days"
            },
            "description": "Number of days to analyze for RFM calculation"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/segments/distribution": {
      "get": {
        "tags": [
          "Segments"
        ],
        "summary": "Get Segment Distribution",
        "description": "Get distribution of users across different segments\nReturns percentage breakdown and counts",
        "operationId": "get_segment_distribution_api_segments_distribution_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SegmentStatsResponse": {
        "properties": {
          "segment_name": {
            "type": "string",
            "title": "Segment Name"
          },
          "user_count": {
            "type": "integer",
            "title": "User Count"
          },
          "avg_ltv": {
            "type": "number",
            "title": "Avg Ltv"
          },
          "activity_level": {
            "type": "string",
            "title": "Activity Level"
          }
        },
        "type": "object",
        "required": [
          "segment_name",
          "user_count",
          "avg_ltv",
          "activity_level"
        ],
        "title": "SegmentStatsResponse"
      },
...
      "UpdateSegmentRequest": {
        "properties": {
          "user_id": {
            "type": "integer",
            "title": "User Id"
          },
          "rfm_group": {
            "type": "string",
            "title": "Rfm Group",
            "description": "RFM segment group"
          },
          "ltv_score": {
            "type": "number",
            "title": "Ltv Score",
            "description": "Lifetime value score",
            "default": 0.0
          },
          "risk_profile": {
            "type": "string",
            "title": "Risk Profile",
            "description": "Risk profile level",
            "default": "MEDIUM"
          }
        },
        "type": "object",
        "required": [
          "user_id",
          "rfm_group"
        ],
        "title": "UpdateSegmentRequest"
      },
...
      "UserSegmentResponse": {
        "properties": {
          "user_id": {
            "type": "integer",
            "title": "User Id"
          },
          "rfm_group": {
            "type": "string",
            "title": "Rfm Group"
          },
          "ltv_score": {
            "type": "number",
            "title": "Ltv Score"
          },
          "risk_profile": {
            "type": "string",
            "title": "Risk Profile"
          },
          "last_updated": {
            "type": "string",
            "format": "date-time",
            "title": "Last Updated"
          }
        },
        "type": "object",
        "required": [
          "user_id",
          "rfm_group",
          "ltv_score",
          "risk_profile",
          "last_updated"
        ],
        "title": "UserSegmentResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      },
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "api/auth/login"
          }
        }
      }
    }
  }
}